Updating namespaces
