--- src/APSIM.Builds/Data/INextGenDbContextGenerator.cs
@@ -1,7 +1,0 @@
-namespace APSIM.Builds.Data
-{
-    public interface INextGenDbContextGenerator
-    {
-        INextGenDbContext GenerateDbContext();
-    }
-}
