+++ src/APSIM.Builds/Data/OldApsim/IOldApsimDbContextGenerator.cs
@@ -1,7 +1,12 @@
-namespace APSIM.Builds.Data.OldApsim
+namespace APSIM.Builds.Data.OldApsim;
+
+/// <summary>
+/// An interface for a class which can construct old apsim DB contexts.
+/// </summary>
+public interface IOldApsimDbContextGenerator
 {
-    public interface IOldApsimDbContextGenerator
-    {
-        IOldApsimDbContext GenerateDbContext();
-    }
+    /// <summary>
+    /// Create an old apsim DB context.
+    /// </summary>
+    IOldApsimDbContext GenerateDbContext();
 }
