+++ src/APSIM.Builds/Data/OldApsim/OldApsimDbContext.cs
@@ -4,24 +4,23 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace APSIM.Builds.Data.OldApsim
+namespace APSIM.Builds.Data.OldApsim;
+
+public class OldApsimDbContext : DbContext, IOldApsimDbContext
 {
-    public class OldApsimDbContext : DbContext, IOldApsimDbContext
+    public DbSet<Build> Builds { get; set; }
+
+    /// <summary>
+    /// Create a new <see cref="OldApsimDbContext"/>.
+    /// </summary>
+    /// <param name="options">DB context builder options.</param>
+    public OldApsimDbContext(DbContextOptions options) : base(options)
     {
-        public DbSet<Build> Builds { get; set; }
+    }
 
-        /// <summary>
-        /// Create a new <see cref="OldApsimDbContext"/>.
-        /// </summary>
-        /// <param name="options">DB context builder options.</param>
-        public OldApsimDbContext(DbContextOptions options) : base(options)
-        {
-        }
-
-        /// <inheritdoc />
-        public Task<int> SaveChangesAsync()
-        {
-            return base.SaveChangesAsync();
-        }
+    /// <inheritdoc />
+    public Task<int> SaveChangesAsync()
+    {
+        return base.SaveChangesAsync();
     }
 }
