+++ src/APSIM.Builds/Models/Platform.cs
@@ -1,12 +1,11 @@
-namespace APSIM.Builds
+namespace APSIM.Builds;
+
+/// <summary>
+/// Platforms supported by APSIM Next Gen.
+/// </summary>
+public enum Platform
 {
-    /// <summary>
-    /// Platforms supported by APSIM Next Gen.
-    /// </summary>
-    public enum Platform
-    {
-        Linux,
-        Windows,
-        MacOS
-    }
+    Linux,
+    Windows,
+    MacOS
 }
