+++ src/APSIM.Builds/Pages/_ViewImports.cshtml
@@ -1,3 +1,2 @@
-﻿@using src
-@namespace src.Pages
+﻿@namespace APSIM.Builds.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
